Revert "[FIXED] use int for millisecond"
This reverts commit 5a9aa4103deb35f386928839210e3afd10e200f9.

+++ glp_loop.c
@@ -9,7 +9,7 @@
 
 struct loop_state {
 	struct glp_clock* clk;
-	uint32_t sleep_time;
+	float sleep_time;
 
 	int fps_game;
 	// time per frame
